pyproject.toml
[tool.poetry]
name = "it-scout-test-task"
version = "0.1.0"
description = ""
authors = ["Khalil <h.sultanoff@yandex.ru>"]
readme = "README.md"
packages = [{ include = "it_scout_test_task" }]

[tool.poetry.dependencies]
python = ">=3.11"
amqp = "5.3.1"
asgiref = "3.8.1"
async-timeout = "5.0.1"
attrs = "25.3.0"
billiard = "4.2.1"
celery = "5.5.2"
certifi = "2025.4.26"
charset-normalizer = "3.4.2"
click = "8.2.1"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
django = "5.2.1"
django-request = "1.7.0"
django-rest-framework = "0.1.0"
djangorestframework = "3.16.0"
drf-spectacular = "0.28.0"
future = "1.0.0"
idna = "3.10"
inflection = "0.5.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2025.4.1"
kombu = "5.5.3"
mccabe = "0.7.0"
mypy-extensions = "1.1.0"
packaging = "25.0"
pathspec = "0.12.1"
pillow = "11.2.1"
platformdirs = "4.3.8"
prompt-toolkit = "3.0.51"
psycopg2-binary = "2.9.10"
pycodestyle = "2.13.0"
pyflakes = "3.3.2"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.1.0"
pyyaml = "6.0.2"
redis = "6.1.0"
referencing = "0.36.2"
requests = "2.32.3"
rpds-py = "0.25.1"
serializers = "0.2.4"
six = "1.17.0"
sqlparse = "0.5.3"
typing-extensions = "4.13.2"
tzdata = "2025.2"
uritemplate = "4.1.1"
urllib3 = "2.4.0"
vine = "5.1.0"
wcwidth = "0.2.13"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
mypy = "^1.15.0"
pytest = "^8.3.5"
types-requests = "^2.32.0.20250515"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
exclude = "migrations,settings.py"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py"]
